+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [9.1.0] - 2018-12-20
+
+### New
+
+- Support for several kinds of data files using [require-data](https://github.com/electerious/require-data)
+- Support for data files returning a function
+- Support for data files returning an async function
+
 ## [9.0.0] - 2018-08-25
 
 ### Changed